pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Daniil Likhachev <d.v.likhachev101@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
tortoise-orm = "^0.20.0"
pydantic-settings = "^2.2.1"
uvicorn = "^0.27.1"
asyncpg = "^0.29.0"
aiogram = "^3.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
